feat: introduce nix flake
Prior to this change, the project could not be built on a standard NixOS
installation.

This change introduces two files:
1. flake.nix, which declares all the necessary system components that
   are required to build and run the various parts of the project; and
2. flake.lock, which pins those components to specific versions to
   ensure reprodicible builds and consistent results.

The Cargo workspace has been selected as the default package and running
the following command (requires Nix 2.5) will place all the project's
binaries in `result/bin`:

    nix build

A devShell is also included, which means that running the following
command will enter the same environment use by nix build with the
addition of cargo-make and mzoon packages.